abstracts[] - add secure_storage - add intl localizations - new file: lib/core/storage/storage.dart - new file: lib/feature/_feature.dart - new file: lib/feature/example/_example.dart - new file: lib/feature/example/data/_date.dart - new file: lib/feature/example/data/models/_models.dart - new file: lib/feature/example/data/repo/_repo.dart - new file: lib/feature/example/pages/_pages.dart - new file: lib/feature/example/state/_state.dart - new file: lib/feature/example/widgets/_widgets.dart - new file: lib/feature/hive/_hive.dart - new file: lib/feature/hive/pages/_pages.dart - new file: lib/feature/hive/pages/hive.dart - new file: lib/feature/secure_storage/secure_storage.dart - new file: lib/generated/intl/messages_all.dart - new file: lib/generated/intl/messages_en.dart - new file: lib/generated/intl/messages_kg.dart - new file: lib/generated/intl/messages_ru.dart - new file: lib/generated/l10n.dart - new file: lib/l10n/intl_en.arb - new file: lib/l10n/intl_kg.arb - new file: lib/l10n/intl_ru.arb - modified: lib/main.dart - modified: pubspec.lock - modified: pubspec.yaml
