+++ lib/feature/hive/pages/_pages.dart
@@ -1,0 +1,1 @@
+export 'hive.dart';
