+++ lib/l10n/intl_en.arb
@@ -1,0 +1,6 @@
+{
+  "title_hiveTest": "Hello HiveTest",
+  "subtitle_hivetest": "HiveTest",
+  "text_delete":"Заметка была удалена"
+  
+}
