+++ lib/l10n/intl_kg.arb
@@ -1,0 +1,1 @@
+{}
