+++ lib/l10n/intl_ru.arb
@@ -1,0 +1,5 @@
+{
+    "title_hiveTest": "Привет HiveTest",
+    "subtitle_hivetest": "HiveTest",
+    "text_delete":"Note has been deleted"
+  }
